backend/requirements.txt
fastapi
uvicorn
python-multipart
yt-dlp
opencv-python-headless
numpy
requests
Pillow
imagehash
sse-starlette
groq
python-dotenv
